requirements.txt
fastapi
uvicorn
fastapi_sqlalchemy
python-dotenv
SQLAlchemy
psycopg2-binary
